Convert price to double in setPrice declaration

+++ src/RoomType.java
@@ -55,7 +55,7 @@ public class RoomType {
         return price;
     }
     
-    public void setPrice(long price) {
+    public void setPrice(double price) {
         this.price = price;
     }
